Fix for Orc Deaths
-The Issue: In MEP version 0.6b, all orcs are given the trait orkish,
the trait in this case gives 20 health and has the following name and
description.
orkish;Creature of Melkor;;;;;;x
orkish_desc;This character belongs to one of the corrupted races,
created by Melkor.;;;;;;x
-In our current version, this trait was changed to -20 health and the
localization was changed to the following:
orkish;At Deaths Door;;;;;;x
orkish_desc;This character is going to die.;;;;;;x
-Furthermore, the trait icon was changed from the "Creature of Melkor"
icon to the syphilitic icon.
-My Fix: I changed the race trait name to orc, restored the 0.6b
settings and icon.  I have left the changes to "orkish" alone because
they are obviously intentional and I don't want to step on someone's
toes.