x11 Release changes...
- Succession for Wights (including the Nazgûl) is now 'Open Elective'.
-
Added an event to re-imprison Morgoth if Manwë banishes or frees him.
-
The Swan Knights, the Rangers of Ithilien, and the Iron Guard are now
vassalized and do not exist when they should not. (Also, removed the
other Rangers of Ithilien). However, they have not been fully checked
for consistency (also, the Grey Company has not been touched yet).
-
Imported and de-vanillified some decision and event files (often
deleting all events) to try and reduce CTD/Asserts. Decisions are now
MEP only.
- Added d_sauron_exile to give Sauron a landless title when he
is not otherwise in power.
- Moved the old forest into a separate duchy
and kingdom
- Created a 'Huorn Defence Force' event for the Old Forest
and Fangorn.
- misc bugfixes