x13
- Fixed some bad prerequisites for dwarven 'city' water-systems and
mines.
- Last Alliance bugfixes, added some events to disband the Last
Allice once the Anti-Mordor war is over.
- Altered purge events so that
rather than the completion being scheduled 10 years in-advance for the
same leader only, purging-provinces are checked annually and completed
if they are unoccupied and it has been 9.9 years or more since the
purging started.
- Added player-only Wilderness-Colonization with
Orcs/Trolls for Evil Overlords.
- Fiddled around with some race trait
events, so elven childhood is applied at game start, and checks are done
more frequently.
- Event targets for many important characters set at
game start.
- Made some edits to the 'Rohan-Isengard Wars' (covering
Gandalf-Saruman and the Battles of the ford of Isen), that should fix
some errors.
- Incorporated hhardy01's Númenórean character changes.
-
The Weather Hills remain part of Arthedian until 9900 (thanks to
Deaghaidh)
- Map Edits: Added the missing province Kor Mumak (thanks to
Esdf20), and passages Rivendell-to-Coldfells, and
Adligwold-to-Thakalgund.
- Validated holy sites (the Elves had 8,
Umaiar, Spiders, and Dragons had None). All 'religions' now have 5 holy
sites.
- Orcs and Trolls can now have consorts.
- Various bugfixes